+++ packages/terra-core-docs/CHANGELOG.md
@@ -6,6 +6,10 @@
   * Updated Arrange component example to use `span` element instead of `div`.
 
 * Added
+  * Added realistic examples for `terra-arrange`.
+  * Added realistic examples for `terra-toggle-section-header`.
+
+* Added
   * Added email validation for `terra-form-field`.
 
 * Updated